+++ js/settings.js
@@ -14,6 +14,8 @@
 //		Service used for website icons: Google (ggl), DuckDuckGo (ddg)
 
 // USER SETTINGS
+// hide_attributes
+//		Hide the attributes a-z, A-Z, 0-9, !@#
 // hide_passwords
 //		Hide passwords by showing them as '*****'
 // hide_usernames
@@ -174,6 +176,7 @@ $(document).ready(function() {
 
 	$('#hide_usernames').prop('checked', (settings.getKey('hide_usernames').toLowerCase() == 'true'));
 	$('#hide_passwords').prop('checked', (settings.getKey('hide_passwords').toLowerCase() == 'true'));
+	$('#hide_attributes').prop('checked', (settings.getKey('hide_attributes').toLowerCase() == 'true'));
 	
 	// save it all
 	$('#savepersonal').click(function() {
@@ -183,6 +186,7 @@ $(document).ready(function() {
 		settings.setUserKey('icons_show', $('#icons_show').is(':checked'));
 		settings.setUserKey('hide_usernames', $('#hide_usernames').is(':checked'));
 		settings.setUserKey('hide_passwords', $('#hide_passwords').is(':checked'));
+		settings.setUserKey('hide_attributes', $('#hide_attributes').is(':checked'));
 
 		// no error, so:
 		if ($('.msg-passwords').text = t('passwords', 'Saving...')) {